Refactor: Move ZeldaObjectData and step_hits
